Rasa_Chatbot(rerank)(4)/requirements.txt
chromadb==0.6.3
fastapi==0.115.11
langchain==0.3.20
langchain_openai==0.3.8
langdetect==1.0.9
pandas==2.2.3
pydantic==2.10.6
python-dotenv==1.0.1
rank_bm25==0.2.2
rasa_sdk==3.11.0
Requests==2.32.3
scikit_learn==1.1.3
sentence_transformers==3.4.1
spacy==3.8.3
uvicorn==0.34.0
